Constructors2
The main purpose of this example is to show that when called, a constructor for the Android class will call upon another constructor. Constructor #1 calls Constructor #2 by using this(boolean, String, String, int, int), because Constructor #2 shares the exact same name as Constructor #1. By inputting the required value types in the parameter field in Constructor #1, it calls and executes Constructor #2.

Now in the main program (outside the Android class) when an instance object of class Android is created, it will automatically be assigned the data values set up by Constructor #2 and defined in #1.